@@ -11,19 +11,20 @@
   },
   "main": "lib/index.js",
   "dependencies": {
-    "@google-cloud/bigquery": "^1.2.0",
     "@slack/client": "^4.2.0",
-    "actions-on-google": "2.0.1",
-    "dialogflow": "^0.3.0",
-    "dialogflow-fulfillment": "0.3.0-beta.2",
-    "firebase-admin": "~5.12.0",
-    "firebase-functions": "^1.0.2",
-    "moment": "^2.22.1",
-    "octokat": "^0.10.0"
+    "firebase-admin": "~11.9.0",
+    "firebase-functions": "^4.4.0",
+    "moment": "^2.29.4",
+    "axios": "^1.4.0"
   },
   "devDependencies": {
-    "tslint": "^5.8.0",
-    "typescript": "^2.8.3"
+    "@typescript-eslint/eslint-plugin": "^5.59.9",
+    "@typescript-eslint/parser": "^5.59.9",
+    "eslint": "^8.42.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-prettier": "^4.2.1",
+    "typescript": "^5.1.3"
   },
   "private": true
 }
